+++ packages/core/create-dlightjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dlightjs
 
+## 0.10.0
+
+### Minor Changes
+
+- feat: upgrade to 0.10.0
+
 ## 0.9.34
 
 ### Patch Changes