+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/dlight
 
+## 0.10.0
+
+### Minor Changes
+
+- feat: upgrade to 0.10.0
+
 ## 0.9.30
 
 ### Patch Changes