+++ packages/ecosystem/decorators/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/decorators
 
+## 0.10.0
+
+### Minor Changes
+
+- feat: upgrade to 0.10.0
+
 ## 0.9.1
 
 ### Patch Changes