+++ packages/ecosystem/markit/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @dlightjs/markit
 
+## 0.10.0
+
+### Minor Changes
+
+- feat: upgrade to 0.10.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/dlight@0.10.0
+
 ## 0.9.38
 
 ### Patch Changes