+++ packages/ecosystem/material-icons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @dlightjs/material-icons
 
+## 0.10.0
+
+### Minor Changes
+
+- feat: upgrade to 0.10.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/decorators@0.10.0
+  - @dlightjs/dlight@0.10.0
+
 ## 0.9.30
 
 ### Patch Changes