+++ packages/ecosystem/transpiler-standalone/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @dlightjs/transpiler-standalone
 
+## 0.10.0
+
+### Minor Changes
+
+- feat: upgrade to 0.10.0
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@0.10.0
+
 ## 0.9.29
 
 ### Patch Changes